requirements.txt
playwright>=1.47.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
requests>=2.32.3
pandas>=2.2.2
tqdm>=4.66.5
pydantic>=2.8.2
python-dotenv>=1.0.1
pdfplumber>=0.11.4
openai>=1.0.0
msal>=1.24.0
